Move `DeanonymizedParticipation` to deployments subsystem.
