+++ typescript-declarations/tests/carp.deployments.core.ts
@@ -16,6 +16,7 @@ import MasterDeviceDeployment = dk.cachet.carp.deployments.application.MasterDev
 import StudyDeploymentStatus = dk.cachet.carp.deployments.application.StudyDeploymentStatus
 import ActiveParticipationInvitation = dk.cachet.carp.deployments.application.users.ActiveParticipationInvitation
 import AssignedMasterDevice = dk.cachet.carp.deployments.application.users.AssignedMasterDevice
+import DeanonymizedParticipation = dk.cachet.carp.deployments.application.users.DeanonymizedParticipation
 import ParticipantData = dk.cachet.carp.deployments.application.users.ParticipantData
 import Participation = dk.cachet.carp.deployments.application.users.Participation
 import StudyInvitation = dk.cachet.carp.deployments.application.users.StudyInvitation
@@ -47,6 +48,8 @@ describe( "carp.deployments.core", () => {
             ActiveParticipationInvitation.Companion,
             new AssignedMasterDevice( exampleDevice, null ),
             AssignedMasterDevice.Companion,
+            new DeanonymizedParticipation( UUID.Companion.randomUUID(), UUID.Companion.randomUUID() ),
+            DeanonymizedParticipation.Companion,
             new ParticipantData( UUID.Companion.randomUUID(), toMap( [] ) ),
             ParticipantData.Companion,
             new Participation( UUID.Companion.randomUUID() ),
